Move heapmanager to dataclay.backend
